+++ packages/core/src/icon/IconWrapper.tsx
@@ -1,9 +1,0 @@
-import styled from '../styled'
-
-export interface IconWrapperProps {
-  color?: string
-}
-
-export const IconWrapper = styled('span')<IconWrapperProps>(({color}) => ({
-  fill: color,
-}))