+++ packages/mobile/styleguide/components/Para.tsx
@@ -1,17 +1,24 @@
 import React, {FC} from 'react'
 
-import {themes, ThemeProvider, Box} from '@qiwi/pijma-core'
+import {Box} from '@qiwi/pijma-core'
 
-import {Paragraph} from '@qiwi/pijma-mobile'
+import {Paragraph, ParagraphProps} from '@qiwi/pijma-mobile'
 
-const ParagraphRenderer: FC = (props) => (
-  <ThemeProvider theme={themes.orange}>
-    <Box my={4}>
-      <Paragraph size="m">
-        {props.children}
-      </Paragraph>
-    </Box>
-  </ThemeProvider>
+const SizeMargin: { [size in NonNullable<ParagraphProps['size']>]: number } = {
+  s: 3,
+  m: 4,
+  l: 5,
+}
+
+const ParagraphRenderer: FC = ({children}) => (
+  <Box
+    my={SizeMargin['m']}
+    css={{'&:first-child': {marginTop: 0}, '&:last-child': {marginBottom: 0}}}
+  >
+    <Paragraph size="m">
+      {children}
+    </Paragraph>
+  </Box>
 )
 
 export default ParagraphRenderer